+++ Core/src/main/java/com/plotsquared/core/events/post/PostPlotChangeOwnerEvent.java
@@ -25,21 +25,54 @@
  */
 package com.plotsquared.core.events.post;
 
-import com.plotsquared.core.events.PlotEvent;
+import com.plotsquared.core.events.PlotPlayerEvent;
+import com.plotsquared.core.player.PlotPlayer;
 import com.plotsquared.core.plot.Plot;
+import org.checkerframework.checker.nullness.qual.Nullable;
+
+import java.util.UUID;
 
 /**
- * Called after a {@link Plot} was deleted.
+ * Called after the owner of a plot was updated.
  */
-public class PlotDeletedEvent extends PlotEvent {
+public class PostPlotChangeOwnerEvent extends PlotPlayerEvent {
+
+    @Nullable
+    private final UUID oldOwner;
 
     /**
-     * Instantiate a new PlotDeleteEvent.
+     * Instantiate a new PlotChangedOwnerEvent.
      *
-     * @param plot The plot which was deleted.
+     * @param initiator The player who executed the owner change.
+     * @param plot      The plot which owner was changed.
+     * @param oldOwner  The previous owner - if present, otherwise {@code null}.
      */
-    public PlotDeletedEvent(final Plot plot) {
-        super(plot);
+    public PostPlotChangeOwnerEvent(final PlotPlayer<?> initiator, final Plot plot, @Nullable UUID oldOwner) {
+        super(initiator, plot);
+        this.oldOwner = oldOwner;
+    }
+
+    /**
+     * @return the old owner of the plot - if present, otherwise {@code null}.
+     */
+    public @Nullable UUID getOldOwner() {
+        return oldOwner;
+    }
+
+    /**
+     * @return {@code true} if the plot had an owner, {@code false} otherwise.
+     * @see #getOldOwner()
+     */
+    public boolean hasOldOwner() {
+        return getOldOwner() != null;
+    }
+
+    /**
+     * @return {@code true} if the plot now has an owner, {@code false} otherwise.
+     * @see Plot#hasOwner()
+     */
+    public boolean hasNewOwner() {
+        return getPlot().hasOwner();
     }
 
 }
