+++ Core/src/main/java/com/plotsquared/core/events/post/PostPlotDeleteEvent.java
@@ -25,25 +25,21 @@
  */
 package com.plotsquared.core.events.post;
 
-import com.plotsquared.core.events.PlotPlayerEvent;
-import com.plotsquared.core.player.PlotPlayer;
+import com.plotsquared.core.events.PlotEvent;
 import com.plotsquared.core.plot.Plot;
 
 /**
- * Called after a plot was chosen for a player by `/plot auto`
- * <br>
- * Called after {@link com.plotsquared.core.events.PlayerAutoPlotEvent}  and only, if no listener cancelled the action.
+ * Called after a {@link Plot} was deleted.
  */
-public class PlayerAutoPlotPostEvent extends PlotPlayerEvent {
+public class PostPlotDeleteEvent extends PlotEvent {
 
     /**
-     * Instantiate a new PlayerAutoPlotPostEvent.
+     * Instantiate a new PlotDeleteEvent.
      *
-     * @param plotPlayer The player who claims a new plot by `/plot auto`.
-     * @param plot       The plot that is assigned to the player.
+     * @param plot The plot which was deleted.
      */
-    public PlayerAutoPlotPostEvent(final PlotPlayer<?> plotPlayer, final Plot plot) {
-        super(plotPlayer, plot);
+    public PostPlotDeleteEvent(final Plot plot) {
+        super(plot);
     }
 
 }
