+++ Core/src/main/java/com/plotsquared/core/events/post/PostPlotMergeEvent.java
@@ -28,51 +28,20 @@ package com.plotsquared.core.events.post;
 import com.plotsquared.core.events.PlotPlayerEvent;
 import com.plotsquared.core.player.PlotPlayer;
 import com.plotsquared.core.plot.Plot;
-import org.checkerframework.checker.nullness.qual.Nullable;
-
-import java.util.UUID;
 
 /**
- * Called after the owner of a plot was updated.
+ * Called after multiple plots were merged with another.
  */
-public class PlotChangedOwnerEvent extends PlotPlayerEvent {
-
-    @Nullable
-    private final UUID oldOwner;
+public class PostPlotMergeEvent extends PlotPlayerEvent {
 
     /**
-     * Instantiate a new PlotChangedOwnerEvent.
+     * Instantiate a new PlotMergedEvent.
      *
-     * @param initiator The player who executed the owner change.
-     * @param plot      The plot which owner was changed.
-     * @param oldOwner  The previous owner - if present, otherwise {@code null}.
+     * @param plotPlayer The {@link PlotPlayer} that initiated the merge.
+     * @param plot       The final merged plot.
      */
-    public PlotChangedOwnerEvent(final PlotPlayer<?> initiator, final Plot plot, @Nullable UUID oldOwner) {
-        super(initiator, plot);
-        this.oldOwner = oldOwner;
-    }
-
-    /**
-     * @return the old owner of the plot - if present, otherwise {@code null}.
-     */
-    public @Nullable UUID getOldOwner() {
-        return oldOwner;
-    }
-
-    /**
-     * @return {@code true} if the plot had an owner, {@code false} otherwise.
-     * @see #getOldOwner()
-     */
-    public boolean hasOldOwner() {
-        return getOldOwner() != null;
-    }
-
-    /**
-     * @return {@code true} if the plot now has an owner, {@code false} otherwise.
-     * @see Plot#hasOwner()
-     */
-    public boolean hasNewOwner() {
-        return getPlot().hasOwner();
+    public PostPlotMergeEvent(final PlotPlayer<?> plotPlayer, final Plot plot) {
+        super(plotPlayer, plot);
     }
 
 }
